edit habit info
