+++ lib/common/common_textfield.dart
@@ -7,31 +7,30 @@ class CommonTextfield extends StatefulWidget {
     this.maxLength,
     this.maxLines = 1,
     this.isTapped,
+    this.initialValue,
   });
 
   final String hintText;
   final int? maxLength;
   final int maxLines;
   final bool? isTapped;
+  final String? initialValue;
 
   @override
   State<CommonTextfield> createState() => _CommonTextfieldState();
 }
 
 class _CommonTextfieldState extends State<CommonTextfield> {
-  final controller = TextEditingController();
-
   @override
   Widget build(BuildContext context) {
     return TextFormField(
+      initialValue: widget.initialValue,
       showCursor: !(widget.isTapped ?? false),
       mouseCursor:
           !(widget.isTapped ?? false) ? SystemMouseCursors.basic : null,
-      controller: controller,
       onChanged: (value) {
         setState(() {});
       },
-      //  onTap: widget.onTap,
       maxLength: widget.maxLength,
       maxLines: widget.maxLines,
       decoration: InputDecoration(
