+++ lib/habits/habits_screen.dart
@@ -71,7 +71,16 @@ class _HabitTilesState extends State<_HabitTiles> {
               destination: ViewHabitScreen(
                 habit: habit,
               ),
-              trailingWidget: const Icon(Icons.flare_outlined),
+              trailingWidget: const Row(
+                mainAxisSize: MainAxisSize.min,
+                children: [
+                  Text('10'),
+                  SizedBox(
+                    width: 5,
+                  ),
+                  Icon(Icons.flare_outlined),
+                ],
+              ),
             ),
           ),
       ],
