+++ pubspec.yaml
@@ -31,6 +31,7 @@ dependencies:
   flutter:
     sdk: flutter
   postgres: ^2.4.3
+  collection: 1.18.0
 
 
   # The following adds the Cupertino Icons font to your application.
